cdk/pyproject.toml
[tool.poetry]
name = "c3l-agent-orchestration"
version = "0.1.0"
description = ""
authors = ["haiyue <haiyue.wang@unisa.edu.au>"]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13.1"
aws-cdk-lib = "^2.173.2"
constructs = "^10.3.0"
cdk-nag = "^2.34.2"
pydantic = "^2.10.2"
cdk-ecr-deployment = "^3.1.4"
aws-cdk-aws-lambda-python-alpha = "^2.179.0a0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pyright = "^1.1.364"
ruff = "^0.5.1"

[tool.pyright]
typeCheckingMode = "standard"
exclude = ["cdk.out"]

[tool.pytest.ini_options]
norecursedirs = "cdk.out/*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
